requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.2
django-allauth==65.1.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
idna==3.10
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
